Add heuristic to the type deduction algorithm
The type deduction algorithm now uses an heuristic
capable of judging which subclass in a hierarchical
lua OOP structure an identifier is set to.
If only generic operations used by several subclasses is used
it will now suggest members of that class, rather than
any class with the given member.