Fix type deduction errors
Previously, autocomplete-lua could suggest types
which are not of the lowest size possible (it would
suggest specialized classes where it should have been
suggesting generalized classes) This has been fixed.

Type deduction could throw errors when trying to deduce
types for variables which table was accessed before it was
defined. Type deduction now handles properly those cases.

Finally, the editor is now a bit more explicit with what type
it has deduced for which variable. It is now shown in the tooltip
the same description for deduced type variables than for variables which
type has been found with the luacompleterc NamedType deduction.